+++ test/TC3/ads-client.test.js
@@ -106,7 +106,8 @@ describe('connection', () => {
       allPlcSymbolsCached: false,
       plcSymbols: {},
       allPlcDataTypesCached: false,
-      plcDataTypes: {}
+      plcDataTypes: {},
+      adsSymbolsUseUtf8: false
     }));
 
     expect(client.metaData.tcSystemState).toStrictEqual({
@@ -123,12 +124,18 @@ describe('connection', () => {
 
     //Note: checking only keys
     expect(Object.keys(client.metaData.plcUploadInfo)).toStrictEqual(Object.keys({
+      version: 0,
       symbolCount: 0,
       symbolLength: 0,
       dataTypeCount: 0,
       dataTypeLength: 0,
-      extraCount: 0,
-      extraLength: 0
+      maxDynamicSymbolCount: 0,
+      dynamicSymbolCount: 0,
+      invalidDynamicSymbolCount: 0,
+      encodingCodePage: 0,
+      flags: 0,
+      flagsStr: [],
+      reserved: Buffer.alloc(0)
     }));
 
     //this didn't work in TC3 PLC 4022.x
@@ -335,12 +342,18 @@ describe('symbols and data types', () => {
 
     //Note: checking only keys
     expect(Object.keys(res)).toStrictEqual(Object.keys({
+      version: 0,
       symbolCount: 0,
       symbolLength: 0,
       dataTypeCount: 0,
       dataTypeLength: 0,
-      extraCount: 0,
-      extraLength: 0
+      maxDynamicSymbolCount: 0,
+      dynamicSymbolCount: 0,
+      invalidDynamicSymbolCount: 0,
+      encodingCodePage: 0,
+      flags: 0,
+      flagsStr: [],
+      reserved: Buffer.alloc(0)
     }));
   });